// src/templates/district-courts.generated.ts
export const JADVAL2_COURTS: DistrictCourtDef[] = [
  {
    code: 'fuq-andijan-andvilfsud',
    regionCode: 'andijan',
    courtTypeCode: 'fuqarolik',
    name: {
      uz_cyrillic: 'Андижон вилоят суди',
      uz_latin: 'Andijon viloyat sudi',
      ru: 'Андижон вилоят суди',
    },
  },
// ...
];